Move util/functions.ts into common/functions.ts
