+++ client/src/main/scripts/common/functions.ts
