+++ client/src/main/scripts/components/BlockReader.tsx
@@ -20,7 +20,7 @@
 import * as React from "react";
 import { FT } from "../java2ts/FT";
 import { Routes } from "../java2ts/Routes";
-import { slugify } from "../utils/functions";
+import { slugify } from "../common/functions";
 import BlockEditor, { TakeDocument } from "./BlockEditor";
 import ReactionContainer from "./ReactionContainer";
