+++ client/src/main/scripts/components/TimelinePreview.tsx
@@ -28,7 +28,7 @@ import {
 } from "../common/CaptionNodes";
 import { FT } from "../java2ts/FT";
 import { Routes } from "../java2ts/Routes";
-import {} from "../utils/functions";
+import {} from "../common/functions";
 import Document, { DocumentEventHandlers } from "./Document";
 import DocumentTextNodeList from "./DocumentTextNodeList";
 import FactHeader, { StickyFactHeader } from "./FactHeader";
