+++ client/src/main/scripts/components/search/VideoResult.tsx
@@ -20,7 +20,7 @@
 import * as React from "react";
 import { FT } from "../../java2ts/FT";
 import { Routes } from "../../java2ts/Routes";
-import { slugify } from "../../utils/functions";
+import { slugify } from "../../common/functions";
 import { convertSecondsToTimestamp } from "../../common/video";
 import { SearchHit } from "./search";
